code/deployment/app/requirements.txt
streamlit 
Pillow